pyproject.toml
[build-system]
requires = ["flit_core>=3.8"]
build-backend = "flit_core.buildapi"

[project]
name = "qwen-image-edit"
version = "0.1.0"
description = "Qwen Image Edit API"
requires-python = ">=3.8"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-dotenv",
  "diffusers",
  "transformers",
  "accelerate",
  "Pillow",
  "python-multipart",
  "requests",
  "torch",
  "torchvision",
  "bitsandbytes",
]
